requirements.txt
mcp==1.27.0
requests==2.33.1
streamlit==1.56.0
sentence-transformers==5.3.0
llama-cpp-python==0.3.20
torchvision==0.26.0
torch==2.11.0














